Randomly assign groups

(Brief description of the function here.)

Usage

assign_groups(sample_size, sample_prob, prob0, prob1, seed)

Arguments

sample_size

total number of people under observation.

sample_prob

a vector of probability weights for obtaining the elements of the vector being sampled.

prob0

vector probability of each possible outcome for the null group

prob1

vector probability of each possible outcome for the intervention group

seed

integer specifying the seed number

Value

list of group assignments